% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_deriv.R
\name{check_deriv}
\alias{check_deriv}
\title{Check gradients are correct}
\usage{
check_deriv(thetas = NULL, mu = 0, Lambda_lt, y, M, method)
}
\arguments{
\item{thetas}{}

\item{mu}{}

\item{Lambda_lt}{}

\item{y}{a T-by-J data matrix}

\item{M}{a length J list with the corresponding covariance matrices}

\item{method}{}
}
\description{
Function to make sure the numerical and anlytical gradients indeed agree
}
