% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_deriv.R
\name{gen_deriv_0}
\alias{gen_deriv_0}
\title{URE}
\usage{
gen_deriv_0(y, M, type)
}
\arguments{
\item{y}{a T-by-J data matrix}

\item{M}{a length J list with the corresponding covariance matrices}

\item{mu}{}
}
\description{
Calculates URE(mu, Lambda) as defined in Kwon (2020)
}
