% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_wrapper.R
\name{gen_obj}
\alias{gen_obj}
\title{Create objective function (URE) for optimizing Lambda}
\usage{
gen_obj(
  y,
  M,
  type,
  W,
  missing_cells,
  O,
  diag_lam,
  centering,
  bounds = NULL,
  Z = list()
)
}
\arguments{
\item{y}{}

\item{M}{}

\item{type}{}

\item{W}{}

\item{missing_cells}{}

\item{O}{}

\item{centering}{}

\item{mu_abs_bounds}{}
}
\description{
Create objective function for optimizing Lambda that takes the centring term
as given
}
