% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_wrapper.R
\name{gen_obj_0}
\alias{gen_obj_0}
\title{Create objective function (URE) for optimizing Lambda}
\usage{
gen_obj_0(y, M, type)
}
\arguments{
\item{y}{}

\item{M}{}

\item{mu}{}
}
\description{
Create objective function for optimizing Lambda that takes the centring term
as given
}
