% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{get_theta_ol}
\alias{get_theta_ol}
\title{...}
\usage{
get_theta_ol(
  thetas,
  y,
  M,
  centering,
  diag_lam = FALSE,
  W = NULL,
  Z = list(),
  n_init_vals,
  all_init_vals = FALSE,
  optim_control
)
}
\arguments{
\item{thetas}{}

\item{y}{}

\item{M}{}

\item{centering}{}

\item{diag_lam}{}

\item{W}{}

\item{n_init_vals}{}

\item{optim_control}{}
}
\description{
..
}
