% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_risk.R
\name{loss}
\alias{loss}
\title{Loss function of thetahat(mu, Lambda) at theta

..}
\usage{
loss(mu, Lambda, y, M, T, theta)
}
\arguments{
\item{mu}{}

\item{Lambda}{}

\item{y}{}

\item{M}{}

\item{T}{}

\item{theta}{T-by-J matrix of true mean vector}
}
\description{
Loss function of thetahat(mu, Lambda) at theta

..
}
