% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_risk.R
\name{loss_diag}
\alias{loss_diag}
\title{Loss fucntion for the independent case

..}
\usage{
loss_diag(mu, Lambda, y, M, T, theta)
}
\arguments{
\item{mu}{}

\item{Lambda}{}

\item{y}{}

\item{M}{}

\item{T}{}

\item{theta}{}
}
\description{
Loss fucntion for the independent case

..
}
