% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_risk.R
\name{loss_generic}
\alias{loss_generic}
\title{Loss function for general estimator

..}
\usage{
loss_generic(theta, thetahat, T)
}
\arguments{
\item{theta}{}

\item{thetahat}{}

\item{T}{}
}
\description{
Loss function for general estimator

..
}
