% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{make_ol_deriv}
\alias{make_ol_deriv}
\title{URE}
\usage{
make_ol_deriv(y, M, thetas)
}
\arguments{
\item{y}{a T-by-J data matrix}

\item{M}{a length J list with the corresponding covariance matrices}
}
\description{
Calculates URE(mu, Lambda) as defined in Kwon (2020)
}
