% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{make_ol_deriv_lt}
\alias{make_ol_deriv_lt}
\title{URE}
\usage{
make_ol_deriv_lt(thetas, y, M, W, missing_cells, O, diag_lam)
}
\arguments{
\item{thetas}{}

\item{y}{}

\item{M}{}

\item{W}{}

\item{missing_cells}{}

\item{O}{}

\item{diag_lam}{}
}
\description{
Calculates URE(mu, Lambda) as defined in Kwon (2020)
}
