% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{make_oracle_cpp_obj}
\alias{make_oracle_cpp_obj}
\title{...}
\usage{
make_oracle_cpp_obj(thetas, mu, y, M, diag)
}
\arguments{
\item{thetas}{}

\item{mu}{}

\item{y}{}

\item{M}{}

\item{diag}{}
}
\description{
..
}
