% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{make_oracle_obj}
\alias{make_oracle_obj}
\title{...}
\usage{
make_oracle_obj(
  thetas,
  y,
  M,
  W,
  missing_cells,
  O,
  diag_lam,
  centering,
  Z = list()
)
}
\arguments{
\item{thetas}{}

\item{y}{}

\item{M}{}

\item{diag_lam}{}

\item{centering}{}
}
\description{
..
}
