% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_wrapper.R
\name{make_URE_cpp_obj}
\alias{make_URE_cpp_obj}
\title{Create objective function (URE) for optimizing Lambda}
\usage{
make_URE_cpp_obj(mu, y, M)
}
\arguments{
\item{mu}{}

\item{y}{}

\item{M}{}
}
\description{
Create objective function for optimizing Lambda that takes the centring term
as given
}
