% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_deriv.R
\name{make_URE_deriv}
\alias{make_URE_deriv}
\title{URE}
\usage{
make_URE_deriv(y, M)
}
\arguments{
\item{y}{a T-by-J data matrix}

\item{M}{a length J list with the corresponding covariance matrices}
}
\description{
Calculates URE(mu, Lambda) as defined in Kwon (2020)
}
