% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebmle.R
\name{nll}
\alias{nll}
\title{Negative log likelihood of the marginal distribution of the data}
\usage{
nll(mu, Lambda, y, M)
}
\arguments{
\item{mu}{}

\item{Lambda}{}

\item{y}{a T-by-J data matrix}

\item{M}{a length J list with the corresponding covariance matrices}
}
\description{
Negative log likelihood of the marginal distribution of the data
}
