% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_deriv.R
\name{nll_deriv_j}
\alias{nll_deriv_j}
\title{Derivative of URE}
\usage{
nll_deriv_j(Lambda, y_j, M_j, mu_j)
}
\arguments{
\item{Lambda}{}

\item{y_j}{}

\item{M_j}{}
}
\description{
Makes the function that calculate the component of URE(mu,
Lambda) that cooresponds to the jth observation ..
}
