% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_wrapper.R
\name{opt_mu_Lambda_URE}
\alias{opt_mu_Lambda_URE}
\title{...}
\usage{
opt_mu_Lambda_URE(Lambda, y, M, mu_abs_bounds)
}
\arguments{
\item{Lambda}{}

\item{y}{}

\item{M}{}

\item{mu_abs_bounds}{}
}
\description{
..
}
