% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{thetahat_diag}
\alias{thetahat_diag}
\title{Shrinkage estimator for theta in the independent case

..}
\usage{
thetahat_diag(mu = 0, Lambda, M, y)
}
\arguments{
\item{mu}{}

\item{Lambda}{}

\item{M}{JT vector of variances}

\item{y}{JT vector of observations}
}
\description{
Shrinkage estimator for theta in the independent case

..
}
