% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{thetahat_j}
\alias{thetahat_j}
\title{Shrinkage estimator for jth observation}
\usage{
thetahat_j(mu, Lambda, Mj, yj)
}
\arguments{
\item{mu}{}

\item{Lambda}{}

\item{Mj}{}

\item{yj}{}
}
\description{
Computes the shrinkage estimator for theta_j

..
}
