% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_estimates.R
\name{URE}
\alias{URE}
\title{URE}
\usage{
URE(mu, Lambda, y, M)
}
\arguments{
\item{mu}{}

\item{Lambda}{}

\item{y}{a T-by-J data matrix}

\item{M}{a length J list with the corresponding covariance matrices}
}
\description{
Calculates URE(mu, Lambda) as defined in Kwon (2020)
}
