% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_estimates.R
\name{URE_cov}
\alias{URE_cov}
\title{URE with covariates}
\usage{
URE_cov(gamma, Lambda, y, M, Z)
}
\arguments{
\item{gamma}{}

\item{Lambda}{}

\item{y}{a T-by-J data matrix}

\item{M}{a length J list with the corresponding covariance matrices}

\item{Z}{}
}
\description{
Calculates URE(Z gamma, Lambda) as defined in Kwon (2020)
}
