% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_deriv.R
\name{URE_deriv_j}
\alias{URE_deriv_j}
\title{make_URE_j}
\usage{
URE_deriv_j(Lambda, y_j, M_j)
}
\arguments{
\item{Lambda}{}

\item{y_j}{}

\item{mu}{}
}
\description{
Makes the function that calculate the component of URE(mu,
Lambda) that cooresponds to the jth observation ..
}
