% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_estimates.R
\name{URE_diag}
\alias{URE_diag}
\title{URE_diag}
\usage{
URE_diag(mu, Lambda, y, M)
}
\arguments{
\item{mu}{location vector to be tuned}

\item{Lambda}{a T vector to be tuned}

\item{y}{a TJ vector of the observations, which is constructed by
concatenating the y_j's (as opposed to concatenating the y_t's)}
}
\description{
Calculates URE(mu, Lambda) as defined in Xie, Kou, and Brown
(2012), but allowed to have different lambdas periodwise
}
