% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_shrink.R
\name{fe_shrink}
\alias{fe_shrink}
\title{Calculate the optiamal fixed effects estimator}
\usage{
fe_shrink(
  y,
  M,
  centering = c("0", "gen", "cov"),
  W = NULL,
  type = c("URE", "EBMLE"),
  init_vals = 1,
  lam_range = c(0.1, 1),
  all_init_val = FALSE,
  use_EBMLE_opt = FALSE,
  use_DEoptim = FALSE,
  print_by_init_val = FALSE
)
}
\arguments{
\item{y}{a T-by-J data matrix, possibly correlated
within each column. If a vector is provided, it is assumed that T=1 and
will be coerced to a 1-by-length(y) matrix.}

\item{M}{length J list of the covaraince matrices.}

\item{centering}{Centering}

\item{type}{}
}
\description{
..
}
