% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{get_theta_ol}
\alias{get_theta_ol}
\title{...}
\usage{
get_theta_ol(thetas, y, M, centering, diag = FALSE, n_init_vals, optim_control)
}
\arguments{
\item{thetas}{}

\item{y}{}

\item{M}{}

\item{centering}{}

\item{diag}{}

\item{n_init_vals}{}
}
\description{
..
}
