% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_estimates.R
\name{URE_diag}
\alias{URE_diag}
\title{URE_diag}
\usage{
URE_diag(mu, Lambda, y, M)
}
\arguments{
\item{mu}{}

\item{Lambda}{}

\item{y}{}

\item{M}{}
}
\description{
Calculates URE(mu, Lambda) as defined in Xie, Kou, and Brown
(2012), but allowed to have different lambdas periodwise
}
