% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_estimates.R
\name{URE_j}
\alias{URE_j}
\title{URE_j}
\usage{
URE_j(mu, Lambda, y_j, M_j)
}
\arguments{
\item{mu}{}

\item{Lambda}{}

\item{y_j}{}

\item{M_j}{}
}
\description{
Makes the function that calculate the component of URE(mu,
Lambda) that cooresponds to the jth observation ..
}
